Move deconvolution scan generation into engine
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
